+++ backend/src/main.rs
@@ -17,13 +17,11 @@ use rand::Rng;
 use sqlx::{postgres::PgListener, Pool, Postgres};
 use std::{
     collections::HashMap,
+    fs::{create_dir_all, DirBuilder},
     net::{IpAddr, Ipv4Addr, SocketAddr},
     time::Duration,
 };
-use tokio::{
-    fs::{create_dir_all, DirBuilder, File},
-    io::AsyncReadExt,
-};
+use tokio::{fs::File, io::AsyncReadExt};
 use uuid::Uuid;
 use windmill_api::HTTP_CLIENT;
 
@@ -34,20 +32,20 @@ use windmill_common::{
     global_settings::{
         BASE_URL_SETTING, BUNFIG_INSTALL_SCOPES_SETTING, CRITICAL_ALERT_MUTE_UI_SETTING,
         CRITICAL_ERROR_CHANNELS_SETTING, CUSTOM_TAGS_SETTING, DEFAULT_TAGS_PER_WORKSPACE_SETTING,
-        DEFAULT_TAGS_WORKSPACES_SETTING, ENV_SETTINGS, EXPOSE_DEBUG_METRICS_SETTING,
-        EXPOSE_METRICS_SETTING, EXTRA_PIP_INDEX_URL_SETTING, HUB_BASE_URL_SETTING, INDEXER_SETTING,
-        INSTANCE_PYTHON_VERSION_SETTING, JOB_DEFAULT_TIMEOUT_SECS_SETTING, JWT_SECRET_SETTING,
-        KEEP_JOB_DIR_SETTING, LICENSE_KEY_SETTING, MONITOR_LOGS_ON_OBJECT_STORE_SETTING,
-        NPM_CONFIG_REGISTRY_SETTING, NUGET_CONFIG_SETTING, OAUTH_SETTING, OTEL_SETTING,
-        PIP_INDEX_URL_SETTING, REQUEST_SIZE_LIMIT_SETTING,
-        REQUIRE_PREEXISTING_USER_FOR_OAUTH_SETTING, RETENTION_PERIOD_SECS_SETTING,
-        SAML_METADATA_SETTING, SCIM_TOKEN_SETTING, SMTP_SETTING, TEAMS_SETTING,
-        TIMEOUT_WAIT_RESULT_SETTING,
+        DEFAULT_TAGS_WORKSPACES_SETTING, EMAIL_DOMAIN_SETTING, ENV_SETTINGS,
+        EXPOSE_DEBUG_METRICS_SETTING, EXPOSE_METRICS_SETTING, EXTRA_PIP_INDEX_URL_SETTING,
+        HUB_BASE_URL_SETTING, INDEXER_SETTING, INSTANCE_PYTHON_VERSION_SETTING,
+        JOB_DEFAULT_TIMEOUT_SECS_SETTING, JWT_SECRET_SETTING, KEEP_JOB_DIR_SETTING,
+        LICENSE_KEY_SETTING, MONITOR_LOGS_ON_OBJECT_STORE_SETTING, NPM_CONFIG_REGISTRY_SETTING,
+        NUGET_CONFIG_SETTING, OAUTH_SETTING, OTEL_SETTING, PIP_INDEX_URL_SETTING,
+        REQUEST_SIZE_LIMIT_SETTING, REQUIRE_PREEXISTING_USER_FOR_OAUTH_SETTING,
+        RETENTION_PERIOD_SECS_SETTING, SAML_METADATA_SETTING, SCIM_TOKEN_SETTING, SMTP_SETTING,
+        TEAMS_SETTING, TIMEOUT_WAIT_RESULT_SETTING,
     },
     scripts::ScriptLang,
     stats_ee::schedule_stats,
     utils::{hostname, rd_string, Mode, GIT_VERSION},
-    worker::{reload_custom_tags_setting, HUB_CACHE_DIR, TMP_DIR, WORKER_GROUP},
+    worker::{reload_custom_tags_setting, HUB_CACHE_DIR, TMP_DIR, TMP_LOGS_DIR, WORKER_GROUP},
     DB, METRICS_ENABLED,
 };
 
@@ -73,7 +71,7 @@ use windmill_worker::{
     DENO_CACHE_DIR_NPM, GO_BIN_CACHE_DIR, GO_CACHE_DIR, LOCK_CACHE_DIR, PIP_CACHE_DIR,
     POWERSHELL_CACHE_DIR, PY310_CACHE_DIR, PY311_CACHE_DIR, PY312_CACHE_DIR, PY313_CACHE_DIR,
     RUST_CACHE_DIR, TAR_PIP_CACHE_DIR, TAR_PY310_CACHE_DIR, TAR_PY311_CACHE_DIR,
-    TAR_PY312_CACHE_DIR, TAR_PY313_CACHE_DIR, TMP_LOGS_DIR, UV_CACHE_DIR,
+    TAR_PY312_CACHE_DIR, TAR_PY313_CACHE_DIR, UV_CACHE_DIR,
 };
 
 use crate::monitor::{
@@ -140,8 +138,8 @@ async fn cache_hub_scripts(file_path: Option<String>) -> anyhow::Result<()> {
         )
     })?;
 
-    create_dir_all(HUB_CACHE_DIR).await?;
-    create_dir_all(BUN_BUNDLE_CACHE_DIR).await?;
+    create_dir_all(HUB_CACHE_DIR)?;
+    create_dir_all(BUN_BUNDLE_CACHE_DIR)?;
 
     for path in paths.values() {
         tracing::info!("Caching hub script at {path}");
@@ -152,7 +150,7 @@ async fn cache_hub_scripts(file_path: Option<String>) -> anyhow::Result<()> {
             .is_some_and(|x| x == &ScriptLang::Deno)
         {
             let job_dir = format!("{}/cache_init/{}", TMP_DIR, Uuid::new_v4());
-            create_dir_all(&job_dir).await?;
+            create_dir_all(&job_dir)?;
             let _ = windmill_worker::generate_deno_lock(
                 &Uuid::nil(),
                 &res.content,
@@ -170,7 +168,7 @@ async fn cache_hub_scripts(file_path: Option<String>) -> anyhow::Result<()> {
         } else if res.language.as_ref().is_some_and(|x| x == &ScriptLang::Bun) {
             let job_id = Uuid::new_v4();
             let job_dir = format!("{}/cache_init/{}", TMP_DIR, job_id);
-            create_dir_all(&job_dir).await?;
+            create_dir_all(&job_dir)?;
             if let Some(lockfile) = res.lockfile {
                 let _ = windmill_worker::prepare_job_dir(&lockfile, &job_dir).await?;
                 let envs = windmill_worker::get_common_bun_proc_envs(None).await;
@@ -505,7 +503,6 @@ Windmill Community Edition {GIT_VERSION}
         DirBuilder::new()
             .recursive(true)
             .create("/tmp/windmill")
-            .await
             .expect("could not create initial server dir");
 
         #[cfg(feature = "tantivy")]
@@ -788,11 +785,12 @@ Windmill Community Edition {GIT_VERSION}
                                                 },
                                                 EXPOSE_METRICS_SETTING  => {
                                                     tracing::info!("Metrics setting changed, restarting");
-                                                    // we wait a bit randomly to avoid having all servers and workers shutdown at same time
-                                                    let rd_delay = rand::rng().random_range(0..40);
-                                                    tokio::time::sleep(Duration::from_secs(rd_delay)).await;
-                                                    if let Err(e) = tx.send(()) {
-                                                        tracing::error!(error = %e, "Could not send killpill to server");
+                                                    send_delayed_killpill(&tx, 40, "metrics setting change").await;
+                                                },
+                                                EMAIL_DOMAIN_SETTING => {
+                                                    tracing::info!("Email domain setting changed");
+                                                    if server_mode {
+                                                        send_delayed_killpill(&tx, 4, "email domain setting change").await;
                                                     }
                                                 },
                                                 EXPOSE_DEBUG_METRICS_SETTING => {
@@ -802,29 +800,17 @@ Windmill Community Edition {GIT_VERSION}
                                                 },
                                                 OTEL_SETTING => {
                                                     tracing::info!("OTEL setting changed, restarting");
-                                                    // we wait a bit randomly to avoid having all servers and workers shutdown at same time
-                                                    let rd_delay = rand::rng().random_range(0..4);
-                                                    tokio::time::sleep(Duration::from_secs(rd_delay)).await;
-                                                    if let Err(e) = tx.send(()) {
-                                                        tracing::error!(error = %e, "Could not send killpill");
-                                                    }
+                                                    send_delayed_killpill(&tx, 4, "OTEL setting change").await;
                                                 },
                                                 REQUEST_SIZE_LIMIT_SETTING => {
                                                     if server_mode {
                                                         tracing::info!("Request limit size change detected, killing server expecting to be restarted");
-                                                        // we wait a bit randomly to avoid having all servers shutdown at same time
-                                                        let rd_delay = rand::rng().random_range(0..4);
-                                                        tokio::time::sleep(Duration::from_secs(rd_delay)).await;
-                                                        if let Err(e) = tx.send(()) {
-                                                            tracing::error!(error = %e, "Could not send killpill to server");
-                                                        }
+                                                        send_delayed_killpill(&tx, 4, "request size limit change").await;
                                                     }
                                                 },
                                                 SAML_METADATA_SETTING => {
                                                     tracing::info!("SAML metadata change detected, killing server expecting to be restarted");
-                                                    if let Err(e) = tx.send(()) {
-                                                        tracing::error!(error = %e, "Could not send killpill to server");
-                                                    }
+                                                    send_delayed_killpill(&tx, 0, "SAML metadata change").await;
                                                 },
                                                 HUB_BASE_URL_SETTING => {
                                                     if let Err(e) = reload_hub_base_url_setting(&db, server_mode).await {
@@ -1047,7 +1033,6 @@ pub async fn run_workers(
         DirBuilder::new()
             .recursive(true)
             .create(x)
-            .await
             .expect("could not create initial worker dir");
     }
 
@@ -1097,4 +1082,19 @@ pub async fn run_workers(
 
     futures::future::try_join_all(handles).await?;
     Ok(())
+}
+
+async fn send_delayed_killpill(
+    tx: &tokio::sync::broadcast::Sender<()>,
+    max_delay_secs: u64,
+    context: &str,
+) {
+    // Random delay to avoid all servers/workers shutting down simultaneously
+    let rd_delay = rand::rng().random_range(0..max_delay_secs);
+    tracing::info!("Scheduling {context} shutdown in {rd_delay}s");
+    tokio::time::sleep(Duration::from_secs(rd_delay)).await;
+
+    if let Err(e) = tx.send(()) {
+        tracing::error!(error = %e, "Could not send killpill for {context}");
+    }
 }